More feature 2.5

+++ login2-v2/app/controllers/comments_controller.rb
@@ -1,7 +1,7 @@
 class CommentsController < ApplicationController
 
 	def index
-		@comments = Post.find(params[:post_id]).comments.all
+		@comments = Post.find(params[:post_id]).comments.order(:created_at).reverse
 		@comment = Comment.new
 		@post = Post.find(params[:post_id])
 		@user = @post.user
@@ -19,6 +19,7 @@ class CommentsController < ApplicationController
 	def create
 	    @post = Post.find(params[:post_id])
 	    @comment = @post.comments.new(comment_params)
+	    @comment.user_id = current_user.id
 	    if (@comment.save)
 	    	redirect_to post_comments_path(@post)
 	    else
@@ -30,10 +31,12 @@ class CommentsController < ApplicationController
   		@post = Post.find(params[:post_id])
   		@comment = Comment.find(params[:id])
   	end
+
   	def update
   		@comment = Comment.find(params[:id])
   		@post = @comment.post
-  		if current_user() && @comment.update(comment_params)
+  		current_user = current_user()
+  		if current_user && validate(current_user, @comment) && @comment.update(comment_params)
   			flash[:notice]= "Update successfully"
   		else
   			flash[:notice]= "May be login or not comment's owner?"
@@ -45,21 +48,23 @@ class CommentsController < ApplicationController
   		current_user = current_user()
   		post = Post.find_by(id: params[:post_id])
   		if current_user
-			comment = Comment.find_by(id: params[:id])
-			if comment # current_user.id == comment.user_id
-				flash[:notice] = "You delete successfully"
-				comment.destroy
-			end
-		else
-			flash[:notice] = "You can't delete this comment!"
-		end
-		redirect_to post_comments_path(post)
+  			comment = Comment.find_by(id: params[:id])
+  			if comment && validate(current_user, comment)
+  				flash[:notice] = "delete successfully"
+  				comment.destroy
+  			else
+          flash[:notice] = "can't delete this comment!"
+        end
+  		else
+        flash[:notice] = "can't delete this comment!"
+  		end
+		  redirect_to post_comments_path(post)
   	end
 
 
   	private 
   	def comment_params
-  		params.require(:comment).permit(:content)
+  		params.require(:comment).permit(:content, :comment_image)
   	end
   	def current_user
   		if (session[:user_id])
